journal endpoints authenticated

+++ src/main/java/com/edigest/journalApp/service/JournalEntryService.java
@@ -6,6 +6,7 @@ import com.edigest.journalApp.entity.User;
 import com.edigest.journalApp.repository.JournalEtryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -29,7 +30,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEtryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception e){
             System.out.println(e);
             throw new RuntimeException("An error occured while saving the entry",e);
@@ -49,11 +50,26 @@ public class JournalEntryService {
         return journalEtryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEtryRepository.deleteById(id);
-    }
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed=false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed= user.getJournalEntries().removeIf(x-> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEtryRepository.deleteById(id);
+            }
+        }catch( Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An Error occured while deleting the Entry",e);
+        }
+        return removed;
+
+        }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
 
 }
